LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-objects/parseAll-invalid-expected.txt


PASS CSSStyleValue.parseAll() with empty property name throws TypeError
PASS CSSStyleValue.parseAll() with unsupported property name throws TypeError
FAIL CSSStyleValue.parseAll() with invalid value for valid property throws TypeError assert_throws_js: function "() => CSSStyleValue.parseAll('width', '10deg')" threw object "SyntaxError: 10deg cannot be parsed as a width" ("SyntaxError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL CSSStyleValue.parseAll() with invalid value for shorthand property throws TypeError assert_throws_js: function "() => CSSStyleValue.parseAll('margin', '10deg')" threw object "SyntaxError: 10deg cannot be parsed as a margin" ("SyntaxError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS CSSStyleValue.parseAll() with invalid value for custom property throws TypeError
